requirements.txt
uvicorn
fastapi
httpie
pydantic
pytest
sqlalchemy
python-dotenv